% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blrdata}
\alias{blrdata}
\title{Coding of 5,000 Speech-Units from of spoken colloquial New Zealand English}
\format{A data frame with 5,000 rows and 5 variables:
\describe{
  \item{ID}{speaker identification}
  \item{Gender}{gender of speaker (Men versus Women)}
  \item{Age}{age of speaker (Young versus Old)}
  \item{EH}{DESCRIPTION NEEDED}
  \item{Ethnicity}{ethnicity of speaker (Maori versus Pakeha)}
  ...
}}
\source{
Data generated from a sample of the New Zealand component of the International Corpus of Engish \url{https://www.ice-corpora.uzh.ch/en.html}
}
\usage{
blrdata
}
\description{
A dataset containing the coding of 5,000 speech-units of New Zealand English for instances of speech-unit final EH, a speaker id as well as the age, gender, and ethnicity of the speaker.
}
\keyword{datasets}
