% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.summary.R
\name{blrm.summary}
\alias{blrm.summary}
\title{Summary Tables for Binomial Logistic Regressions}
\usage{
blrm.summary(x, a, accuracy)
}
\arguments{
\item{x}{A glm object of family "binomial".}

\item{a}{A lrm object.}

\item{accuracy}{The accuracy of the model predictions.}
}
\description{
This function produces summary tables for fixed-effects binomial logistic regressions by extracting the relevent information from a glm and an lrm object.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{binomial}
\keyword{function}
\keyword{logistic}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
