% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfadata}
\alias{cfadata}
\title{Coding of 250 cases of a fictional data set}
\format{A data frame with 5,000 rows and 5 variables:
\describe{
  \item{Variety}{vareity of English (American versus British)}
  \item{Gender}{gender of speaker (Men versus Women)}
  \item{Age}{age of speaker (Young versus Old)}
  \item{Class}{socioeconomic class of speaker (Middle versus Working)}
  \item{Frequency}{freqeuncy counts of an unnamed linguistic feature}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
cfadata
}
\description{
A dataset containing the 250 cases representing speakers with certain sociodemographic features and their use of an unnamed linguistic feature.
}
\keyword{datasets}
