% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citdata}
\alias{citdata}
\title{Fictional Data set of 303 Users and Non-Users of Discourse Like.}
\format{A data frame with 303 rows and 4 variables:
\describe{
  \item{Gender}{Gender of speaker (Man versus Woman)}
  \item{Age}{Age of speaker (Young versus Old)}
  \item{SocialStatus}{Social status of speakers (High versus Low)}
  \item{LikeUser}{Speaker is a user of discourse like (1: yes versus 0:no)}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
citdata
}
\description{
A fictional dataset containing the coding of 303 speakers that are either users of discourse like or not and their social chracteristics  EH, a speaker id as well as the age, gender, and ethnicity of the speaker.
}
\keyword{datasets}
