% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcR.r
\name{ConcR}
\alias{ConcR}
\title{Concordancing Loaded Files}
\usage{
ConcR(strings, pattern, context)
}
\arguments{
\item{strings}{A vector of character strings.}

\item{pattern}{A charater string representing the pattern that the files are searched for which can contain regular expressions.}

\item{context}{A numeric value which determines how man characters are displayed before and after the search results.}
}
\description{
This function creates keyword-in-context (KWIC) displays from a vector of cahracter strings.
}
\examples{
\dontrun{
ConcR(corpus, "word", 20)
}
}
\keyword{Character}
\keyword{Concordancing,}
\keyword{Function.}
\keyword{Srtings,}
