% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countdata}
\alias{countdata}
\title{Fictional Data of 500 Conversations Coded for Number of Fillers (uhm).}
\format{A data frame with 303 rows and 6 variables:
\describe{
  \item{ID}{Identification of the conversation.}
  \item{Trial}{Integer showing whether the conversation was recorded on day 1, 2, or 3.}
  \item{Language}{Factor representing the language in which the conversation took place (English, German, Russian, Mandarin).}
  \item{Gender}{Factor representing the gender of speakers (Man versus Woman).}
  \item{UHM}{Frequency of pauses filled with uhm.}
  \item{Shots}{Number of shots a speaker drank before the converstion.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
countdata
}
\description{
A fictional data set representing 500 conversations coded for the number of pauses filled with "uhm" along with information about the conversation and speakers.
}
\keyword{datasets}
