% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflm.r
\name{eflm}
\alias{eflm}
\title{Extract Effect Size (Pearson's rho) from Linear Regressions}
\usage{
eflm(x)
}
\arguments{
\item{x}{A glm object of family "gaussian".}
}
\description{
This function returns the Pearson's rho values from fixed-effects linear regressions from a glm object.
}
\examples{
\dontrun{
model.glmer = glmer(depvar ~ indepvar + (1|ranvar), data = data, family = "gaussian")
eflm(model.glmer)
}
}
\keyword{effect}
\keyword{function}
\keyword{linear}
\keyword{pearson's}
\keyword{regression,}
\keyword{rho,}
\keyword{size,}
