% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpR.r
\name{expR}
\alias{expR}
\title{Function for Estimating Expected Type II Error Rates}
\usage{
expR(x)
}
\arguments{
\item{x}{A glm object of family "gaussian".}
}
\description{
This function returns the likelihood od type II errors (beta errors) for multiple linear regressions given their sample size.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{II}
\keyword{beta}
\keyword{error,}
\keyword{linear}
\keyword{multiple}
\keyword{regression}
\keyword{regression,}
\keyword{type}
