% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmmdata}
\alias{lmmdata}
\title{537 Texts coded for their Relative Number of Prepositions.}
\format{A data frame with 537 rows and 5 variables:
\describe{
  \item{Date}{Year of composition.}
  \item{Genre}{Genre of the text.}
  \item{Text}{Author of the text.}
  \item{Prepositions}{Frequency of prepositions per 1,000 words.}
  \item{Region}{Region in which the text was written.}
  ...
}}
\source{
Data generated from a sample of the New Zealand component of the Penn Parsed Corpora of Historical English \url{https://www.ling.upenn.edu/hist-corpora/}
}
\usage{
lmmdata
}
\description{
Data set representing 537 texts coded for the relative frequency of prepositions in them along with information about the texts and their writers.
}
\keyword{datasets}
