% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mblrdata}
\alias{mblrdata}
\title{Fictional Data Set Representing 2000 Speech-Units Coded for Speech-Unit Final Discourse Like.}
\format{A data frame with 2000 rows and 6 variables:
\describe{
  \item{ID}{Anonymous identification sequence for the speaker.}
  \item{Gender}{Factor representing the gender of speakers (Men versus Women).}
  \item{Age}{Factor representing the age of speakers (Old versus Young).}
  \item{ConversationType}{Factor representing whether the conversation took place between interlocutors of the same gender or not (MixedGender versus SameGender).}
  \item{Priming}{Factor representing if another instance of speech-unit final like had occured shortly before (NoPrime versus Prime).}
  \item{SUFlike}{Interger representing whether the speech-unit ended with discourse like or not.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
mblrdata
}
\description{
Fictional data set representing 2000 speech-units coded for whether they ended with speech-unit final like along with information about the sentences and their speakers.
}
\keyword{datasets}
