% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFittingSummarySWSD.R
\name{mdl.fttng.swsd}
\alias{mdl.fttng.swsd}
\title{Summary Tables for the Step-Wise Step-Down Model Fitting Process of Mixed-Effects Binomial Logistic Regression Models}
\usage{
mdl.fttng.swsd(mdlcmp)
}
\arguments{
\item{mdlcmp}{A list of model comparisons that result from anova results (e.g. m0m1: anova(m0, m1, test = "Chi))}
}
\description{
This function produces summary tables for fixed-effects binomial logistic regressions by extracting the relevent information from a glm and an lrm object.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{binomial}
\keyword{function}
\keyword{logistic}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
