% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFittingSummarySWSU.r
\name{mdlfttngswsu}
\alias{mdlfttngswsu}
\title{Summary Tables for Step_Wise Step_Down Model Fitting}
\usage{
mdlfttngswsu(mdlcmp)
}
\arguments{
\item{mdlcmp}{is a list of model comparisons of which individual elements are created by anova calls.}
}
\description{
This function produces summary tables for step-wise step-down model fitting effects binomial logistic regressions by extracting the relevent information from a glm and an lrm object.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{fitting,}
\keyword{function}
\keyword{model}
\keyword{step-down,}
\keyword{step-wise}
\keyword{summary}
\keyword{table,}
