% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlinrsummary.r
\name{mlinrsummary}
\alias{mlinrsummary}
\title{Summary Tables for Multiple Linear Regressions}
\usage{
mlinrsummary(mlr, glm, ia = T)
}
\arguments{
\item{mlr}{A lm object created with the "lm" function.}

\item{glm}{A glm object of family "gaussian" created with the "glm" function..}

\item{ia}{If ia = T (default) the table will show all columns of the summary table.}
}
\description{
This function produces summary tables for fixed-effects multiple linear regressions by extracting the relevent information from a glm and an lm object.
}
\examples{
\dontrun{
Example code will come later!
}
}
\keyword{function}
\keyword{linear}
\keyword{multiple}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
