% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlinr.summary.r, R/mlr.summary.R
\name{mlr.summary}
\alias{mlr.summary}
\title{Summary Tables for Multiple Linear Regressions}
\usage{
mlr.summary(mlr, glm, ia = T)

mlr.summary(mlr, glm, ia = T)
}
\arguments{
\item{mlr}{A linear regression model object created with the lm function.}

\item{glm}{A linear regression model glm object of family "gaussian"  created with the glm function.}

\item{ia}{If ia = T (default) the table will show all columns of the summary table.}

\item{x}{A glm object of family "binomial".}

\item{a}{A lrm object.}
}
\description{
This function produces summary tables for multiple linear regressions by extracting the relevent information from a glm and a lm object.

This function produces summary tables for fixed-effects multiple linear regressions by extracting the relevent information from a glm and an lm object.
}
\examples{
\dontrun{
Example code will come later!
}
\dontrun{
Example code will come later!
}
}
\seealso{



}
\keyword{function}
\keyword{linear}
\keyword{multiple}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
