% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlrdata}
\alias{mlrdata}
\title{Data Set Representing the Price of 100 Presents.}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{status}{Factor representing the relationship status of the buyer (Relationship versus Single).}
  \item{attraction}{Factor representing if the buyer is attracted to the Receiver of the Present (Intersted versus NotIntersted).}
  \item{money}{Numeric variable encoding the price of the present.}
  ...
}}
\source{
Data taken from Field, Andy, Miles, Jeremy, and Field, Zoe. 2012. Discovering Statistics Using R, SAGE. for training purposes.

Fictional data generated for training purposes.
}
\usage{
mlrdata

mlrdata
}
\description{
Fictional data set taken from Field, Andy, Miles, Jeremy, and Field, Zoe. 2012. Discovering Statistics Using R. SAGE which represents the prices of 100 presents along with information about the buyer of the present.

Fictional data representing 6 speakers and providing information about their age and whether they are users of discourse like or not.
}
\keyword{datasets}
