% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple ggplots in one Graphics Window}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Further options passed to \pkg{ggplot} and \pkg{grid}}

\item{plotlist}{THIS NEEDS TO BE DOCUMENTED}

\item{file}{THIS NEEDS TO BE DOCUMENTED}

\item{cols}{THIS NEEDS TO BE DOCUMENTED}

\item{layout}{THIS NEEDS TO BE DOCUMENTED}
}
\description{
This function allows the display of multiple ggplots in a single graphics window.
}
\examples{
\dontrun{
#p1 <- ggplot(data, aes(var1, var2)) +
#geom_point()
#p2 <- ggplot(data, aes(var1, var2)) +
#geom_line()
#multiplot(p1, p2, cols = 1)
}
}
\keyword{function}
\keyword{ggplot2,}
\keyword{multiple}
\keyword{plots,}
