% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordinaldata}
\alias{ordinaldata}
\title{Fictional Data Set Representing 400 observations of Likert Responses.}
\format{A data frame with 400 rows and 3 variables:
\describe{
  \item{apply}{DESCRIPTION NEEDED}
  \item{pared}{Factor representing whether at least one of the student's parents has a graduate degree (1: yes versus 0: no).}
  \item{public}{Factor representing whether the student attended a public or private high-school (1: public versus 0: private).}
  \item{gpa}{Numeric variable representing the student’s grade point average.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
ordinaldata
}
\description{
Fictional data taken from the UCLA Institute for Digital Research and Education (https://stats.idre.ucla.edu/r/dae/ordinal-logistic-regression/) representing 400 responses to Likert items and information about the background of the responders.
}
\keyword{datasets}
