% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posdata}
\alias{posdata}
\title{Fictional Data Set Representing 200 observations of Pauses in Conversations.}
\format{A data frame with 200 rows and 4 variables:
\describe{
  \item{Id}{Identification code of the conversation.}
  \item{Pauses}{Number of pauses in a conversation.}
  \item{Language}{Factor representing the language of the conversation (English versus Russian versus German).}
  \item{Alcohol}{Numeric variable representing mililiters of alcohol drunk by the speaker.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
posdata
}
\description{
Fictional data representing 200 conversations and the number of pauses in the conversations along with information about the conversation.
}
\keyword{datasets}
