% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posdata2}
\alias{posdata2}
\title{Fictional Data Set Representing 500 observations of Pauses in Conversations.}
\format{A data frame with 500 rows and 8 variables:
\describe{
  \item{ID}{Identification code of the conversation.}
  \item{Trial}{Integer showing whether the conversation was recorded on day 1, 2, or 3.}
  \item{Language}{Factor representing the language in which the conversation took place (English, German, Russian, Mandarin).}
  \item{Gender}{Factor representing the gender of speakers (Man versus Woman).}
  \item{UHM}{Frequency of pauses filled with uhm.}
  \item{Noise}{Factor representing if music was playing in the background (with versus Without).}
  \item{Interloctutor}{*THIS IS MISPELLED IN THE CODE* Factor representing the gender of the interlocutor (OtherGender versus SameGender).}
  \item{Shots}{Number of shots a speaker drank before the converstion.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
posdata2
}
\description{
Fictional data representing 500 conversations and the number of pauses in the conversations along with information about the conversation.
}
\keyword{datasets}
