% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudor2.R
\name{pseudor2}
\alias{pseudor2}
\title{Pseudo R2 for Binomial Logistic Mixed-Effects Models}
\usage{
pseudor2(f, d)
}
\arguments{
\item{f}{A vector of two character strings (e.g. c("dv ~ a + b", "(1 + a | Subject)")). The first element is the fixed-effect part of the regression formula (e.g. "DP ~ Predictor1 + Predictor1") and the second part is the random effect structure (e.g. "(1 | Subject)").}

\item{d}{The data set on which the model was fit}
}
\description{
This function pseudo R2 (Nagelkerke and Cox & Snell) for mixed-effects binomial logistic regression models.
}
\examples{
\dontrun{
#model.glmermodel = glmer(depvar ~ indepvar + (1|ranvar), data = data, family = "binomial")
#pseudor2(c("depvar ~ indepvar", "(1|ranvar)"), data)
}
}
\keyword{R2,}
\keyword{binomial}
\keyword{function}
\keyword{logistic}
\keyword{mixed-effects}
\keyword{pseudo}
\keyword{regression,}
