% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.summary.tb.r
\name{slr.summary}
\alias{slr.summary}
\title{Summary Tables for Simple Linear Regressions}
\usage{
slr.summary(x)
}
\arguments{
\item{x}{A simple linear regressions by extracting the relevent information from a regression object created with the lm function.}
}
\description{
This function produces summary tables for simple linear regressions by extracting the relevent information from a regression object created with the lm function.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{function}
\keyword{linear}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
