% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slrdata2}
\alias{slrdata2}
\title{Fictional Data Set Representing 500 observations of Pauses in Conversations.}
\format{A data frame with 60 rows and 2 variables:
\describe{
  \item{Group}{Factor representing the group in which the score was achieved.}
  \item{Score}{Numerc value representing the test score.}
  ...
}}
\source{
Fictional data generated for training purposes.
}
\usage{
slrdata2
}
\description{
Fictional data representing 60 test scores of two groups of students.
}
\keyword{datasets}
