% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrsummary.r
\name{slrsummary}
\alias{slrsummary}
\title{Summary Tables for Simple Linear Regressions}
\usage{
slrsummary(x)
}
\arguments{
\item{x}{A simple linear regressions by extracting the relevent information from a regression object created with the lm function.}
}
\description{
This function produces summary tables for simple linear regressions by extracting the relevent information from a regression object created with the lm function.
}
\examples{
\dontrun{
model.lm <- lm(depvar ~ indepvar, data = data)
slrsummary(model.lm)
}
}
\keyword{function}
\keyword{linear}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
