% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeMLR.r
\name{smplesz}
\alias{smplesz}
\title{Function which evaluates Sample Size}
\usage{
smplesz(x)
}
\arguments{
\item{x}{A multiple linear regression model object.}
}
\description{
This function evaluates whether the sample size is sufficient for a regression analysis based on  is sufficient based on Green, Samuel B. (1991) How many subjects does it take to do a regression analysis. Multivariate Behavioral Research. 26(3): 499-510. summary tables for simple linear regressions by extracting the relevent information from a regression object created with the lm function.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{function}
\keyword{linear}
\keyword{regression,}
\keyword{sample}
\keyword{size,}
