% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsummary.r
\name{blrmsummary}
\alias{blrmsummary}
\title{Summary Tables for Binomial Logistic Regressions}
\usage{
blrmsummary(x, a, accuracy)
}
\arguments{
\item{x}{A glm object of family "binomial".}

\item{a}{A lrm object.}

\item{accuracy}{The accuracy of the model predictions.}
}
\description{
This function produces summary tables for fixed-effects binomial logistic regressions by extracting the relevent information from a glm and an lrm object.
}
\examples{
\dontrun{
model.glm = glm(depvar ~ indepvar, data = data, family = binomial)
model.lrm = lrm(depvar ~ indepvar, data = data, x = T, y = T)
data$predicted <- predict(model.glm, data)
caret::confusionMatrix(data$depvar, data$predict)
# inspect accuracy and include its numeric value in the blrmsummary function call
blrmsummary(model.glm, model.lrm, 80)
}
}
\keyword{binomial}
\keyword{function}
\keyword{logistic}
\keyword{regression,}
\keyword{summary}
\keyword{table,}
