% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlfttngswsu.r
\name{mdlfttngswsu}
\alias{mdlfttngswsu}
\title{Summary Tables for Step_Wise Step_Down Model Fitting}
\usage{
mdlfttngswsu(mdlcmp)
}
\arguments{
\item{mdlcmp}{is a list of model comparisons of which individual elements are created by anova calls.}
}
\description{
This function produces summary tables for step-wise step-down model fitting effects binomial logistic regressions by extracting the relevent information from a glm and an lrm object.
}
\examples{
\dontrun{
m0 <-glm(depvar ~ 1, data = data, family = "binomial")
m1 = glm(depvar ~ indepvar1, data = data, family = "binomial")
m2 = glm(depvar ~ indepvar + inpepvar2, data = data, family = "binomial")
m1m0 = anova(m1, m0, test = "Chi)
m2m1 = anova(m2, m1, test = "Chi)
mdlfttngswsd(m1m0, m2m1)
}
}
\keyword{fitting,}
\keyword{function}
\keyword{model}
\keyword{step-down,}
\keyword{step-wise}
\keyword{summary}
\keyword{table,}
