% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.r
\name{multiplot}
\alias{multiplot}
\title{Multiple ggplots in one Graphics Window}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{x}{A vector of numbers representing the panel which a plot will occupy.}
}
\description{
This function allows the display of multiple ggplots in a single graphics window.
}
\examples{
\dontrun{
Example code will come later!
}
}
\seealso{

}
\keyword{function}
\keyword{ggplot2,}
\keyword{multiple}
\keyword{plots,}
